% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRRCalc.R
\name{IRR}
\alias{IRR}
\title{Incident Rate Ratio Calculator}
\usage{
IRR(count1, volume1, count2, volume2)
}
\arguments{
\item{count1}{Incident count for the first sample.}

\item{volume1}{Volume or standardizing value for the first sample.}

\item{count2}{Incident count for the second sample.}

\item{volume}{2 Volume or standardizing value for the second sample.}
}
\value{
Returns a data frame with the Incident Rate Ratio (IRR) and the lower and upper 95% confidence bound.
}
\description{
This simple function calculates the ratio of two incident rates along with a 95% confidence bound for the ratio.
}
\references{
Algorithm based on "Comparing two incident rate ratios,” https://stats.stackexchange.com/questions/297994/comparing-two-incidence-rate-ratios
}
